% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewinteract.R
\name{get_rule}
\alias{get_rule}
\title{Internal function for taking rules from an rpart stump}
\usage{
get_rule(tree, keepcomplement = TRUE)
}
\arguments{
\item{tree}{rpart tree.}

\item{keepcomplement}{logical or character vector of length one.
If \code{TRUE}, two perfectly collinear rules will be returned. 
If \code{"right"}, the rule indicating membership of the right 
node of \code{tree} will be returned. If \code{"left"}, the rule 
indicating membership of the left node of \code{tree} will be 
returned.}
}
\description{
\code{get_rule} returns a character vector of rules, 
derived from the rpart tree supplied as input.
}
