% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewinteract.R
\name{get_rule_mat}
\alias{get_rule_mat}
\title{Internal function for generating 0-1 dummy coded rule matrix}
\usage{
get_rule_mat(rules, views, firstview, secondview)
}
\arguments{
\item{rules}{character vector with rules. Rules are conditions of input 
variable values.}

\item{views}{named list containing matrices with different views of the
same observations, containing the variables specfied \code{rules}.}

\item{firstview}{the index of the matrix in \code{views} that was
used for creating the first split.}

\item{secondview}{the index of the matrix in \code{views} that was used 
for creating the second and third splits.}
}
\description{
\code{get_rule_mat} generates a matrix of 0-1 coded dummary indicator 
variables, based on a character vector of rules and matrices of input
variables.
}
