% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewinteract.R
\name{get_y_learn_logistic}
\alias{get_y_learn_logistic}
\title{Internal function for updating response in gradient boosting}
\usage{
get_y_learn_logistic(eta, y)
}
\arguments{
\item{eta}{weighted predictions up to the current iteration.}

\item{y}{pseudo response from the first iteration.}
}
\description{
\code{get_y_learn_logistic} returns the pseudoresponse, based
on a current weightd sum of linear predictors (eta) and the 
original response (y).
}
